+++ lib/scraper_utils.rb
@@ -12,6 +12,7 @@ require "scraper_utils/debug_utils"
 require "scraper_utils/log_utils"
 require "scraper_utils/randomize_utils"
 require "scraper_utils/scheduler"
+require "scraper_utils/spec_support"
 
 # Mechanize utilities
 require "scraper_utils/mechanize_actions"